requirements.txt
python-binance==1.0.19
pandas==2.1.4
pandas_ta==0.3.14b0
pytz==2023.3.post1